requirements.txt
flask==3.1.0
openai==1.58.1
anthropic==0.40.0
pyyaml==6.0.2
requests==2.32.3
click==8.1.7
